requirements.txt
Flask==3.0.2
pandas==2.2.1
openpyxl==3.1.2
clickhouse-connect==0.7.1
python-dotenv==1.0.1
